Move events search into its own component
