+++ graylog2-web-interface/src/components/events/events/Events.css
@@ -2,10 +2,6 @@
     margin-top: 10px;
 }
 
-:local(.inline) {
-    display: inline-block;
-}
-
 :local(.priority) {
     font-size: 1.5em;
     vertical-align: top;
