+++ graylog2-web-interface/src/components/events/events/EventsSearchBar.css
@@ -1,0 +1,3 @@
+:local(.inline) {
+    display: inline-block;
+}
